+++ linqex/__init__.py
@@ -1,2 +1,2 @@
 from linqex.linq import *
-from linqex.builds import *
+from linqex.build import *