+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # **Change Log**
 All notable changes to this project will be documented in this file.
 
+## **[1.4] - 19.04.2023**
+Added classes generic structure
+### Added
+ * Added 'EnumerableBase' and 'Enumerable' classes generic structure.
+ * Added 'EnumerableBase().Accumulate', 'EnumerableBase().RemoveAll' and 'EnumerableBase.Dict().OfTypeByKey' methods into the 'Enumerable Base' class
+ * Added 'Enumerable().Accumulate', 'Enumerable().RemoveAll' and 'Enumerable.Dict().OfTypeByKey' methods in the 'Enumerable' class
+### Remove
+ * Removed 'Enumerable().Map' methods in the 'Enumerable' class
+
+<br>
+
 ## **[1.3] - 27.03.2023**
 Most methods found in the C# Linq framework have been transferred to the python-linqex project and additional modifications have been made to the classes. (build.py and linq.py edited, newly added methods tested)
 ### Add